+++ src/neo4j/_async/driver.py
@@ -312,10 +312,10 @@ class AsyncGraphDatabase:
                     #     'Routing parameters are not supported with scheme '
                     #     '"bolt". Given URI "{}".'.format(uri)
                     # )
-                return cls.bolt_driver(parsed.netloc, **config)
+                return cls._bolt_driver(parsed.netloc, **config)
             # else driver_type == DRIVER_NEO4J
             routing_context = parse_routing_context(parsed.query)
-            return cls.neo4j_driver(
+            return cls._neo4j_driver(
                 parsed.netloc, routing_context=routing_context, **config
             )
 
@@ -404,7 +404,7 @@ class AsyncGraphDatabase:
         )
 
     @classmethod
-    def bolt_driver(cls, target, **config):
+    def _bolt_driver(cls, target, **config):
         """
         Create a direct driver.
 
@@ -417,36 +417,28 @@ class AsyncGraphDatabase:
         )
 
         try:
-            return AsyncBoltDriver.open(target, **config)
+            return AsyncBoltDriver._open(target, **config)
         except (BoltHandshakeError, BoltSecurityError) as error:
             from ..exceptions import ServiceUnavailable
 
             raise ServiceUnavailable(str(error)) from error
 
     @classmethod
-    def neo4j_driver(cls, *targets, routing_context=None, **config):
+    def _neo4j_driver(cls, target, routing_context=None, **config):
         """
         Create a routing driver.
 
         Create a driver for routing-capable Neo4j service access
         that uses socket I/O and thread-based concurrency.
         """
-        # TODO: 6.0 - adjust signature to only take one target
-        if len(targets) > 1:
-            deprecation_warn(
-                "Creating a routing driver with multiple targets is "
-                "deprecated. The driver only uses the first target anyway. "
-                "The method signature will change in a future release.",
-            )
-
         from .._exceptions import (
             BoltHandshakeError,
             BoltSecurityError,
         )
 
         try:
-            return AsyncNeo4jDriver.open(
-                *targets, routing_context=routing_context, **config
+            return AsyncNeo4jDriver._open(
+                target, routing_context=routing_context, **config
             )
         except (BoltHandshakeError, BoltSecurityError) as error:
             from ..exceptions import ServiceUnavailable
@@ -455,10 +447,9 @@ class AsyncGraphDatabase:
 
 
 class _Direct:
-    # TODO: 6.0 - those attributes should be private
-    default_host = "localhost"
-    default_port = 7687
-    default_target = ":"
+    _default_host = "localhost"
+    _default_port = 7687
+    _default_target = ":"
 
     def __init__(self, address):
         self._address = address
@@ -468,22 +459,21 @@ class _Direct:
         return self._address
 
     @classmethod
-    def parse_target(cls, target):
+    def _parse_target(cls, target):
         """Parse a target string to produce an address."""
         if not target:
-            target = cls.default_target
+            target = cls._default_target
         return Address.parse(
             target,
-            default_host=cls.default_host,
-            default_port=cls.default_port,
+            default_host=cls._default_host,
+            default_port=cls._default_port,
         )
 
 
 class _Routing:
-    # TODO: 6.0 - those attributes should be private
-    default_host = "localhost"
-    default_port = 7687
-    default_targets = ": :17601 :17687"
+    _default_host = "localhost"
+    _default_port = 7687
+    _default_targets = ": :17601 :17687"
 
     def __init__(self, initial_addresses):
         self._initial_addresses = initial_addresses
@@ -493,15 +483,15 @@ class _Routing:
         return self._initial_addresses
 
     @classmethod
-    def parse_targets(cls, *targets):
+    def _parse_targets(cls, *targets):
         """Parse a sequence of target strings to produce an address list."""
         targets = " ".join(targets)
         if not targets:
-            targets = cls.default_targets
+            targets = cls._default_targets
         return Address.parse_list(
             targets,
-            default_host=cls.default_host,
-            default_port=cls.default_port,
+            default_host=cls._default_host,
+            default_port=cls._default_port,
         )
 
 
@@ -1331,10 +1321,10 @@ class AsyncBoltDriver(_Direct, AsyncDriver):
     """
 
     @classmethod
-    def open(cls, target, **config):
+    def _open(cls, target, **config):
         from .io import AsyncBoltPool
 
-        address = cls.parse_target(target)
+        address = cls._parse_target(target)
         pool_config, default_workspace_config = Config.consume_chain(
             config, AsyncPoolConfig, WorkspaceConfig
         )
@@ -1365,10 +1355,10 @@ class AsyncNeo4jDriver(_Routing, AsyncDriver):
     """
 
     @classmethod
-    def open(cls, *targets, routing_context=None, **config):
+    def _open(cls, *targets, routing_context=None, **config):
         from .io import AsyncNeo4jPool
 
-        addresses = cls.parse_targets(*targets)
+        addresses = cls._parse_targets(*targets)
         pool_config, default_workspace_config = Config.consume_chain(
             config, AsyncPoolConfig, WorkspaceConfig
         )